Use Kahn's algorithm for topological sort
I believe the previous algorithm had an exponential time complexity.

Consider a graph that is a chain of "diamonds":

         *-> B
        /      \
... -> A        *-> D -> ...
        \      /
         *-> C

Each time node D is processed `walk_mut` (assuming that A is not yet
"visited", i.e. processed completely), node A would be added to the
queue twice, and node D once.

Since this is a BFS, the amount of nodes processed at each layer of the
search would grow exponentially until the end of the chain is reached.

This commit also reverses the order of the resulting schedule - not sure
why it was reversed initially.